add automatic connection to secure-mode
