+++ pkg/squashctl/types.go
@@ -5,7 +5,6 @@ import (
 
 	v1 "github.com/solo-io/squash/pkg/api/v1"
 	"github.com/solo-io/squash/pkg/config"
-	"github.com/solo-io/squash/pkg/kscmd"
 	"k8s.io/client-go/kubernetes"
 )
 
@@ -22,8 +21,8 @@ type Options struct {
 	ctx          context.Context
 	Wait         Wait
 
-	Squash  config.Squash
-	Debugee kscmd.Debugee
+	Squash      config.Squash
+	DebugTarget config.DebugTarget
 
 	DeployOptions DeployOptions
